% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egret_flap.R
\name{egret.flap}
\alias{egret.flap}
\title{Creates an EGRET object}
\usage{
egret.flap(df, metadata, pseudotimes)
}
\arguments{
\item{df}{an expression data frame with with cells as columns and features as rows}

\item{metadata}{Additional cell-level metadata to add to the egret object. Used for visualizations, and may include both descriptive and numerical variables.
Should be a data.frame where the rows are cell names and the columns are
additional metadata fields. Row names in the metadata need to match all
column names of the counts data.frame; NA values are allowed.}

\item{pseudotimes}{a vector providing pseudotime values for each cell (calculated externally).}
}
\value{
egret.obj
}
\description{
Creates an EGRET object； takes as input an expression counts table (one feature per row, one cell per column),
metadata associated with each cell (e.g. batch, sample origin, treatment, etc.),
and a vector of pseudotime values (single value per cell).
}
