% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egret_fly.R
\name{egret.fly}
\alias{egret.fly}
\title{Set flight. Calculates gene expression variation with respect to reference trajectory}
\usage{
egret.fly(
  egret,
  steps = 10,
  minCellCount = 3,
  min.detect = 0.1,
  min.pval.cutoff = 0.05
)
}
\arguments{
\item{egret}{An egret object}

\item{steps}{Number of pseudotime steps to bin cells into (integer). Larger values will lead to more finely delinated pseudotime steps, but may also introduce  excessive noise caused by technical variation in sequencing in small numbers of cells. As an approximate guide, the Rice rule (2*cube root of cell count) for histogram binning may be helpful; 1000 cells corresponds to 20 bins. Default value = 10.}

\item{minCellCount}{Minimum number of cells required to define a discrete pseudotime bin. As cells in the same bin are merged together to generate an average expression profile, it is recommended for each bin to include several cells to limit technical noise. Sensible values may range from 3-100s based on the detection sensitivty of the underlying single-cell library construction method and known dropout rate. Default value = 3.}

\item{min.detect}{Minimum detection rate cutoff for a feature to be considered in these analyses. Introducing too many rarely detected genes may introduce technical noise. Default value of 0.1 filters out features expressed in less than 10\% of all cells.}

\item{min.pval.cutoff}{Minimium cutoff p-value for differential expression analysis. Stricter or looser thresholds set the number of DEGs stored. Default t-test p-value = 0.05. Recommended to set a looser threshold for initial calculation, and modify downstream for specific visualizations.}
}
\value{
an egret object
}
\description{
Fly calculates GEX changes with respect to reference trajectory order in terms of three seperate metrics, encompassing number of differentially expressed genes #' between each pseudotime step, mean expression profile similarity between each pseduotime step, and expression enthalpy between each pseudotime step. Calculated values are
stored in associated slots of the Egret object returned as output.
}
