% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCorrelatedGenes.R
\name{FindCorrelatedGenes}
\alias{FindCorrelatedGenes}
\title{Find trajectory-correlated genes}
\usage{
FindCorrelatedGenes(egret, sigDet = 0.2, sigExp = 0.1, sigVar = 1)
}
\arguments{
\item{egret}{An Egret object}

\item{sigDet}{Cutoff for detection rate of a feature across the dataset being considered. For tarjectory-associated genes,
higher detection rates are advisable for picking out genes for more reliable validation. Default = 0.2 (20\% of all cells)}

\item{sigExp}{Cutoff for average detection rate of a feature across the dataset being considered. Genes with higher average detection are likely to be more easily detected.
Default cutoff =0.1 (log2count)}

\item{sigVar}{Variance cutoff for expression variance of a feature across an entire dataset. Genes with too low variance are likely to have low-magnitude expression changes #' which may be difficult to validate. Calculated as standard deviation of log2Count.}
}
\value{
cor.out
}
\description{
Find genes with expression changes tightly correlated with a given trajectory.
These include genes with relatively stable changes (increase or decrease) that reflect the dominant trajectory changes.
}
