% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindUnstableGenes.R
\name{FindUnstableGenes}
\alias{FindUnstableGenes}
\title{Find trajectory-unstable genes}
\usage{
FindUnstableGenes(egret, sigPval = 0.01, sigDiff = 0.25)
}
\arguments{
\item{egret}{An Egret object}

\item{sigPval}{p-value cutoff for expression varaible feature detection.}

\item{sigDiff}{Threshold for maximum expression vairance between any two bins (log2 fold change)}
}
\value{
t1
}
\description{
Find trajectory-unstable genes that show multiple changes in variation but which may not be neccesarily associated with pseudotime progression.
}
