% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMonocle2.R
\name{RunMonocle2}
\alias{RunMonocle2}
\title{Use monocle to generate trajectory}
\usage{
RunMonocle2(mono, varGenes, reverse = F)
}
\arguments{
\item{mono}{a monocle2 CellDataSet object}

\item{varGenes}{a vector of feature ids used for ordering cells. Must be a subset of all features detected.}

\item{reverse}{Whether to reverse the beginning and end points of the learned biological process. Helpful for cases where the trajectory returned is oriented opposite to
actual biology.}
}
\value{
an updated CellDataSet object, in which phenoData contains values for State and Pseudotime for each cell
}
\description{
Use monocle2 DDRTree algorithm to generate trajectory and perform pseudotime assignments.
}
