% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat2Monocle.R
\name{seurat2Monocle}
\alias{seurat2Monocle}
\title{Make a CellDataSet object from a Seurat object}
\usage{
seurat2Monocle(sc, lowerDetectionLimit = 0.01)
}
\arguments{
\item{sc}{a Seurat object}

\item{lowerDetectionLimit}{the minimum expression level that consistitutes true expression}
}
\value{
Returns a CellDataSet object used in monocle
}
\description{
Make a CellDataSet object used in monocle from a Seurat object to proceed to trajectory analysis
}
