% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratBasic.R
\name{SeuratBasic}
\alias{SeuratBasic}
\title{Make a Seurat object from a data frame.}
\usage{
SeuratBasic(
  sc,
  nfeatures = 3000,
  dims = 1:30,
  Feature_low = 500,
  Feature_high = 10000
)
}
\arguments{
\item{sc}{a data frame with cells as columns and features as rows}

\item{nfeatures}{Number of features to select as top variable features}

\item{dims}{dimensions to use as input features}

\item{Feature_low}{Include cells where at least this many features are detected.}

\item{Feature_high}{Exclude cells where more than this many features are detected.}
}
\value{
Returns a Seurat object containing a UMAP representation
}
\description{
Make a Seurat object and process through dimension reduction steps from a data frame. Includes a very basic data filtering and dimension selection option.
Useful for a cursory analysis. It is highly recommended that single-cell is not cleaned through this function, but rather through careful consideration using dedicated
single-cell data processing packages (e.g. Seurat, SingleCellExperiment).
}
