% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soar_cellstream.R
\name{Soar_cellstream}
\alias{Soar_cellstream}
\title{Plot the number of cells in each time bin}
\usage{
Soar_cellstream(egret, metadata = "CellType")
}
\arguments{
\item{egret}{an EGRET object}

\item{metadata}{A factor classifier of cell metadata (e.g. cell type, sample, condition, etc.)}
}
\value{
a plot
}
\description{
Stream plot showing number of cells belonging to
each group within each time bin and the relative composition changes between bins. Useful as an approximate guide of changes, but more precise calculations of compositional
changes should be performed using more rigorous statistical methods and visualized in linearized plots to avoid possible distortions.
}
