% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soar_DEGs.R
\name{Soar_DEGs}
\alias{Soar_DEGs}
\title{number of DEGs between each bin}
\usage{
Soar_DEGs(egret, sigPval = 0.01, sigDiff = 0.25)
}
\arguments{
\item{egret}{An Egret object}

\item{sigPval}{p-value cutoff used to judge DEG signficance}

\item{sigDiff}{Expression fold change (log2FC) cutoff used to determine DEG significance}
}
\value{
a plot
}
\description{
number of DEGs between each bin calculated using pairwise comparsion across each two given pseudotime bins.
}
