% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soar_GEXtimebin.R
\name{Soar_GEXtimebin}
\alias{Soar_GEXtimebin}
\title{View Expression of given genes over discrete pseudotime bins}
\usage{
Soar_GEXtimebin(egret, pick.features, smooth.span = 0.75, jit.width = 0.5)
}
\arguments{
\item{egret}{An Egret object}

\item{pick.features}{A character vector of feature(s) contained in the gene expression data.}

\item{smooth.span}{Value controlling the smoothness of the loess span for curve fitting.}

\item{jit.width}{Value controlling jitter width. Increase jitter to help prevent overplotting at certain points.}
}
\value{
null
}
\description{
View expression of given genes over discrete pseudotime bins. Useful for manually picked genes (e.g. known markers or functional complexes).
potential visualization for trajectory-stable reference and trajectory-unstable noise genes as well.
}
