% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soar_Mountain_DEGs.R
\name{Soar_Mountain_DEGs}
\alias{Soar_Mountain_DEGs}
\title{Downstream visualizations of DEGs in each bin}
\usage{
Soar_Mountain_DEGs(
  egret,
  sigPval = 0.01,
  sigDiff = 0.25,
  l.low = 0,
  l.mid = 250,
  l.high = 1500,
  c.low = "blue",
  c.mid = "white",
  c.high = "red"
)
}
\arguments{
\item{egret}{An egret object}

\item{sigPval}{p-value cutoff for DEG analysis}

\item{sigDiff}{logFC cutoff for DEG analysis}

\item{l.low}{Minimum value for bins (ideally this value is not greater than the minimum value in any bin).}

\item{l.mid}{Medium value for bins. Set to define color blending of figure legend.}

\item{l.high}{Maximum value for bins.}

\item{c.low}{Low color for bins}

\item{c.mid}{Middle color for bins}

\item{c.high}{High color for bins}
}
\value{
null
}
\description{
Downstream visualizations of DEGs across each pseudotime bin rendered as a pyramid of all possible combinations. Users can modify colors and color range for
optimal visualization of the range of DEGs between bins.
}
