% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soar_Mountain_GEXCor.R
\name{Soar_Mountain_GEXCor}
\alias{Soar_Mountain_GEXCor}
\title{Pyramid plot showing gene expression correlation matrix}
\usage{
Soar_Mountain_GEXCor(
  egret,
  l.low = 0.92,
  l.mid = 0.97,
  l.high = 1,
  c.low = "blue",
  c.mid = "white",
  c.high = "red"
)
}
\arguments{
\item{egret}{An Egret object}

\item{l.low}{Minimum value for bins (ideally this value is not greater than the minimum value in any bin).}

\item{l.mid}{Medium value for bins. Set to define color blending of figure legend.}

\item{l.high}{Maximum value for bins.}

\item{c.low}{Low color for bins}

\item{c.mid}{Middle color for bins}

\item{c.high}{High color for bins}
}
\value{
null
}
\description{
Downstream visualizations of average gene expression correlation across each pseudotime bin rendered as a pyramid of all possible combinations.
Users can modify colors and color range for optimal visualization. Correlation calculated using Pearson's R and expressed as R value.
}
