% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soar_Mountain_StepChange.R
\name{Soar_Mountain_StepChange}
\alias{Soar_Mountain_StepChange}
\title{Pyramid plot showing gene expression enthalpy change across each bin}
\usage{
Soar_Mountain_StepChange(
  egret,
  l.low = 0,
  l.mid = 0.25,
  l.high = 0.5,
  c.low = "blue",
  c.mid = "white",
  c.high = "red"
)
}
\arguments{
\item{egret}{An Egret object}

\item{l.low}{Minimum value for bins (ideally this value is not greater than the minimum value in any bin).}

\item{l.mid}{Medium value for bins. Set to define color blending of figure legend.}

\item{l.high}{Maximum value for bins.}

\item{c.low}{Low color for bins}

\item{c.mid}{Middle color for bins}

\item{c.high}{High color for bins}
}
\value{
null
}
\description{
Downstream visualizations of the overall gene expression difference across each pseudotime bin rendered as a pyramid of all possible combinations.
Users can modify colors and color range for optimal visualization. Correlation calculated as the sum of all variation in gene expression. May be useful to conceptualize
as being similar to thermodynamic reaction enthalpy.
}
