% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soar_StepDist.R
\name{Soar_StepDist}
\alias{Soar_StepDist}
\title{Downstream visualizations: Expression step change (enthalpy) across each bin}
\usage{
Soar_StepDist(egret, Ethreshold = 0.5)
}
\arguments{
\item{egret}{An egret object}

\item{Ethreshold}{a threshold value depicted as a line in the plot. Useful for pointing out potential points of discontinuity.
Outliers can be selected acocrding to user choice. As one possibility, values over 3 median absolute deviations from the median can be useful as a guide.}
}
\value{
null
}
\description{
Downstream visualizations step change across each bin, showing the nearest pseudotime bins.
Useful as a more direct visualization compared to the pyramid plots.
}
